test(workflow): fold orphaned node-context propagation test into run_node_test.go
This change deletes workflow/node_context.go, which left
node_context_propagation_test.go named after a source file that no longer
exists. The test really exercises RunNode propagation of context to a
dynamic-node child, so move it (renamed to
TestRunNode_DynamicChildRecoversOwnContext) together with its fnNode helper
into run_node_test.go and drop the orphaned file.

No behavior change.
